Updated git submodules
Project: clamp master b7bdc481784a08d81f71055264f03b9bb3443576

Wait for database with mariadb connector

With the change from Mysql to MariaDB connector
maxReconnect parameters doesn't work anymore.
We then need to use retriesAllDown instead but
this is doing retries as fast as possible with
no wait between tries so need an arbitrary high
value. With maxReconnect it was waiting a time
exponentially growing so a low value was enough.

Change-Id: I7d03baa69bbee46d82a403f573d9f492f9720a4c
Signed-off-by: eh552t <eh552t@intl.att.com>
Issue-ID: CLAMP-42