Update patch set 1
Patch Set 1:

When you make changes to a submodule, you can commit and review those changes as normal. When you back out into the "superproject" and do a 'git status', git says there are changes in the submodule. I added the submodule to the commit, which was probably not the right thing to do. Should I abandon and try again w/o adding submodule changes?

Patch-set: 1
Label: Code-Review=0