Updated git submodules
Project: appc/deployment master e7dfc10e3cd873085b39ad97f526d41321c5c8c6

Change to openstack conn properties

Adding openstack connection properties so that the iaas
adapter will have a valid endpoint of some kind and will
not cause startup delays and failures.

Change-Id: I817236d87602f3c5f7afb2116eb37facff903ac9
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-id: APPC-235


Removing sdnc resource assignment from start

The sdnc resource assignment provider is failing to start,
plus it is a component that we are not using. This change
allows us to modify the sdnc start script so that component
does not start.

Change-Id: Ie48345aa7834393e08070dd6a022e7cc1a4cc5d8
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-id: CCSDK-102


New version of the aaiclient.properties

This is a new version of the properties file. Modified version
of the file that sdnc uses.

Change-Id: Ib927ece901654ab0cbd86d46e1d3434439768439
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-id: APPC-136