Updated git submodules
Project: so master 7b501309195e1ef87e86a8fcc6775ff10708987c

Merge "MSOCommonBPMN flows are not being deployed"

MSOCommonBPMN flows are not being deployed

Flows in the MSOCommonBPMN jar, which is embedded in the
MSOInfrastructureBPMN war, are not being deployed.

MSOCommonBPMN needs a processes.xml file.

I'm also fixing a problem with CustomE2EGetService.bpmn, which claims
it is the GenericGetService flow.  This creates a conflict with the
real GenericGetService flow.

Issue: SO-199
Change-Id: I1ec9d4ad079abf65b5c276b625b46a86805a15e6
Signed-off-by: Rob Daugherty <rd472p@att.com>