Updated git submodules
Project: so master a10e358bb0d385b22cbd7b38b069401f61206359

Build support for Camunda EE Cockpit

The community edition Camunda Cockpit does not have the
capability to inspect a running or completed process.
This makes it useless for debugging purposes.

If a user has an enterprise license, it should be
possible to build SO using the enterprise version of
Camunda.

This commit modifies the SO poms so that the camunda
version may be specified on the maven command line, as
follows:

-Dcamunda.version=7.7.3-ee -Dcamunda.bpm.webapp.artifact=camunda-webapp-ee-jboss-standalone

A maven repository containing the enterprise artifacts
(in this case, 7.7.3-ee) would also need to be provided
at build time in the maven settings.xml configuration.

Issue: SO-216
Change-Id: Idff63ce70475299924333b37aa51f8d963adf568
Signed-off-by: Rob Daugherty <rd472p@att.com>