Updated git submodules
Project: so master a582a03358c8a56a21922db0edbc8f5bc8bcd224

Merge "Flows fail to update status in requests DB"

Flows fail to update status in requests DB

The fallout and completion handlers are not updating
the request status in the mso_requests database.

The cause is a missing URN mapping that defines the
service endpoint for these updates.

Issue: SO-230
Change-Id: I5b4ff6e9ca2cde6605b3ed7cb7b086041111d31c
Signed-off-by: Rob Daugherty <rd472p@att.com>