Update patch set 1
Patch Set 1:

I'm updating the patchset to pull only 1 depth what is very good here. I think the doc could skip the submodules part. Or it's better to ask for a recursive clone as Jenkins does.

My main point is we can't ensure how the user clones the repository. Then Pulling the submodules is mandatory (or checking that it's has been done). It costs nothing even if it has already be done (e.g. in Jenkins jobs).

Regarding the doc, I precise it's useless to ask installing requirements as they're installed via tox in virtualenv.

The next two commands could be enough.
$ git clone http://gerrit.onap.org/r/doc
$ cd tox && tox

Patch-set: 1
Label: Code-Review=0