Updated git submodules
Project: so master 104aefb5d0fce75ee5c5ec4f4de1a6954925b5b6

Query AAI for ARs on delete

Delete-vcpe code expected allotted-resources to be included within the
SI response from AAI.  That turns out not to be the case.  As a result,
the code for delete-vcpe was modified to query each AR found in the SI
to get it's full info.

Change-Id: I893fb2fcf50a32335bbd68d1aaa84c8d747b14f0
Issue-Id: SO-325
Signed-off-by: Jim Hahn <jrh3@att.com>