Update patch set 1
Patch Set 1: Code-Review-1

This is a good topic to discuss, but the week of the project sign-off is not a good time to be changing the top level document structure. 
let us revisit this in the context of Beijing. We need a good description of onboarding from the perspective of the VNF Provider. VNFRQTS project has repo to handle use cases  descriptions like this and could be a good place to maintain this content going forward. Suggest we create JIRA ticket to include this in VNFRQTS Beijing project Plan.

Patch-set: 1
Reviewer: Gerrit User 1579 <1579@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1