Updated git submodules
Project: so master 812d981957b98e8a10135198e87ee84319f5d7e0

Merge "Fix init errors in WorkflowMessageAdapter"

Fix init errors in WorkflowMessageAdapter

When the workflow message adapter comes up, it throws a
FileNotFoundException because cadi.properties is absent.
This file is related to AAF, which is not in scope for
the Amsterdam release.

Issue: SO-341
Change-Id: I27239e215a523f8535d49d1888acc72deb35b270
Signed-off-by: Rob Daugherty <rd472p@att.com>