Updated git submodules
Project: so master 4af0ffe291ec78310085745e3273d170e3943b3f

Merge "Force source to always be included in JSON"

Force source to always be included in JSON

Modified RequestInfo to pass source to JSON, even when it has the
default value.  Ordinarily, this is unnecessary when the default
value is null, but in this case, the default value is not null,
thus other code to which the JSON is passed expects it to be
populated.

Change-Id: I53fde35b8d46f69234fb51f1cceedb31cedc3335
Issue-Id: SO-340
Signed-off-by: Jim Hahn <jrh3@att.com>