Updated git submodules
Project: appc master e7db07433274257a7b792d1b3cf0da0026fd1d6e

Correcting or skipping tests

These tests seem to be designed to timeout and fail. The
resulting exception is either caught or put as an expectation of
the test. The tests do not actually test anything and are causing
30 minutes of build delay.

Change-Id: I177839c0f08f2c4bd8895e0626f49476b4ed7e85
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-ID: APPC-276