Updated git submodules
Project: appc master 781b1a6df324419c846c84ea983c18fc8362bfd3

Third part of onap rename

This part of the commit changes the folder structure on all other
folders of appc.

Change-Id: I8acfa11cdfcdcd36be0e137245d1dd7324f1abd3
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-ID: APPC-13


Second part of onap rename

This is the second commit of the rename. The folder structure
is renamed for appc-adapters and appc-config in this commit.

Change-Id: Iaa2b8c937ff1ca1b5d1178128961fb115ee65d9b
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-ID: APPC-13