Updated git submodules
Project: sdc master 5fbddfd71bb83ab666a29478825124517d12db9b

fix licensing issues

in the project we are required to have a license text in the root dir.
we have it there and it references both AT&T and Amdocs.
other license were removed and and the text in the package json changed to onap.

Change-Id: I1584d3c56df3287d02f899d45f18ddc3f32d082c
Issue-ID: SDC-737
Signed-off-by: Michael Lando <ml636r@att.com>