Updated git submodules
Project: integration master 8fa101a3dd2edf8e432b9447a886cfb583100fb0

Merge "Add kolla internal vip to no_proxy"

Add kolla internal vip to no_proxy

The internal vip value is calculated at deploy time when
bringing up the OpenStack VM, and isn't able to be included
in the proxy values in the Vagrantfile. Because of this, we
need to add it to no_proxy after it's calculated so that users
and other services can access it in a proxy environment.

Signed-off-by: Nathaniel Potter <nathaniel.potter@intel.com>
Issue-ID: INT-383
Change-Id: Ifed792a4f7e6f4f5a227848486e8165be4a30fc7