Updated git submodules
Project: aai/sparky-be master 10f3bd738258130fa1558dd8ae8ba2a1342993ab

Fixed a curruption that occurs on startup

The start script was adding the 2 extra lines right after the last
character in the file.  The extra line in the file will ensure that it
will be added at the end.  Also, the start script now ensures that only 1
copy of these added attributes exist in the file.

Issue-ID: AAI-717
Change-Id: I5deb2d9e74d48f3c4d2351582f05468590d0b72e
Signed-off-by: Paquette, Francis <francis.paquette@amdocs.com>