Updated git submodules
Project: oom master 07dc7db266ecfb88a783a94fac5c54e2092bda53

Fix readiness-check exception

The problem is that the script tried to
iterate on statuses of pod containers.
when some pod is in state "pending" or
some other case when there is no status,
the statuses is None.
Therefore python throws exception when trying to iterate None,
which is non-iterable.
Issue-ID: OOM-514
Signed-off-by: BorislavG <Borislav.Glozman@amdocs.com>
Change-Id: If57b8a632a83489fddb12373e4047396f5fa08a3