Updated git submodules
Project: policy/engine master 9f19ef6e4c962160eb844da5f505be2fa3901a69

Merge "Upgrade to latest elasticsearch"

Upgrade to latest elasticsearch

Will not clear any CLM issue but this for uses a version that is
using the latest lucene-queryparser. It is noted in the CVE for
lucene-queryparser that elasticsearch is not affected by the
issue. We will upgrade to keep up with the latest code.

Issue-ID: POLICY-507
Change-Id: I8580dbc704c48245faf298304b8aec76c9e773c4
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>