Updated git submodules
Project: oom master 6a84df34eabbdb1ca12ec005c9978082013eb7f9

Add ONAP Quickstart README for Bejing

This is largely a duplication of what will be in
the offical readthedocs documentation. It is put
here now as part of the README to provide immediate
help to those trying to deploy ONAP using an all
Helm deployment mechanism.

The README will later be amended to simply point to
the offical docs once it has been published for Beijing.

Change-Id: I5e3ec279e19617912d28d482e7818c01620d8945
Issue-ID: OOM-827
Signed-off-by: Mike Elliott <mike.elliott@amdocs.com>