Updated git submodules
Project: so master b5723530e99da12b7dc5004ca4de2d7bd85deef1

Merge "sdc-tosca:1.3.3 depends on a SNAPSHOT artifact"

sdc-tosca:1.3.3 depends on a SNAPSHOT artifact

We have to exclude jtosca:1.3.4-SNAPSHOT and import jtosca:1.3.4.
This is really a defect in SDC, but we have to work around it for M4.

Change-Id: I4f204da9c5f3f601ae9a832203439bef7554079d
Issue-ID: SO-536
Signed-off-by: Rob Daugherty <rd472p@att.com>