Updated git submodules
Project: ccsdk/storage/pgaas master bd5129dbad044907bd1818b2f8cda3c60da9b3e8

remove most of this code for R2 Beijing

This package is not being used in R2 Beijing because of the move to Kubernetes with OOM.

If you need this package, use the R1 Amsterdam release packages.

I'm not removing the repo though because it is likely that we will need to add things in R3 Casablanca.

Change-Id: I40a46886cfa2f3803f97918ad5cea149f5fcf696
Signed-off-by: Hansen, Tony (th1395) <th1395@att.com>
Issue-ID: CCSDK-231