Updated git submodules
Project: integration/devtool master efe77c01bdceac8902b91f20cae6c5670db33a84

Fix all-in-one vagrant machine name

Currently when inputting the command for all_in_one, vagrant
can't find a machine defined for that name because it's looking
for a machine called all-in-one. This patch updates the name of
the VM in run.sh to all-in-one.

Change-Id: I599d68c91e4c6e614aa884fe7c305b64bbf88e91
Signed-off-by: Nathaniel Potter <nathaniel.potter@intel.com>
Issue-ID: INT-460