Updated git submodules
Project: music master 193f937a5df3cbe476428e25601d47496b629820

Merge "Pin jnr-unixsocket for Docker plugin to 0.13"

Pin jnr-unixsocket for Docker plugin to 0.13

The current verison of jnr-unixsocket is 0.12 from the Docker
plugin however in order to be able the images for ARM64, it needs
the newer version which has support for it.

There is work upstream to switch the version to 0.18[0] however
it seems that this is still pending, so we can pin it for now until
that merges.

[0]: https://github.com/fabric8io/docker-maven-plugin/pull/770

Change-Id: Ia69b1b1807e79cbc50bd12dbb7dd7f9f3d60cfc9
Issue-ID: CIMAN-154
Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>