Updated git submodules
Project: so master a69ded6458098401f16bd554567d6184ead449b5

Move pnfDefaultTimeout to dmaap.properties

Re: https://gerrit.onap.org/r/#/c/48031/1

SO is failing health check because pnfCheckInputs
is not found in dmaap.properties, and
pnf.properties is not being read, since
applicationContext.xml has two
PropertyPlaceholderConfigurer attributes, but
it's only using one.

Moved defaultTimeout to dmaap.properties, where
applicationContext.xml will be able to read it,
and renamed it to pnfDefaultTimeout to make it
clearer what it is.

Change-Id: I6490de472fbf4875e53204f4248c6325787c8af5
Issue-ID: SO-506
Signed-off-by: Arthur Martella <am153x@att.com>