Updated git submodules
Project: so master ef58615371f366eb8415f69b73ad5e0d52910a11

Fix OOF serviceResourceId bug

- Works around a bug in OOF
where OOF does not return a serviceResourceId
with homing solutions. SO uses serviceResourceId
to match a homing deman to a homing request.
This patch addes resourceModuleName as another
way to match a demand to a request, as currently
OOF needs this to be unique.

- Changes SO homing processing of key cloudRegionId to
key locationId as this is now the key used for cloudRegionId
in solutions sent back from OOF

Issue-ID: OPTFRA-249
Change-Id: I8e5d9244a227ea4d3d061a758f202144bac270ce
Signed-off-by: Marcus G K Williams <marcus.williams@intel.com>