Updated git submodules
Project: aai/esr-server master 6f31082c9081f9e453b7e279f9591cc11eeb9d92

Merge "Pin jnr-unixsocket for Docker plugin to 0.13"

Pin jnr-unixsocket for Docker plugin to 0.13

The current verison of jnr-unixsocket is 0.12 from the Docker
plugin however in order to be able the images for ARM64, it needs
the newer version which has support for it.

There is work upstream to switch the version to 0.18[0] however
it seems that this is still pending, so we can pin it for now until
that merges.

[0]: https://github.com/fabric8io/docker-maven-plugin/pull/770

Change-Id: I5340a2adad30f0d8cef22b9e67c21e1f36883c14
Issue-ID: CIMAN-154
Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>