Updated git submodules
Project: aai/esr-gui master 54a7c358f8d900ed03a9139b91864882c60d9028

Merge "Pin jnr-unixsocket for Docker plugin to 0.13"

Pin jnr-unixsocket for Docker plugin to 0.13

The current verison of jnr-unixsocket is 0.12 from the Docker
plugin however in order to be able the images for ARM64, it needs
the newer version which has support for it.

There is work upstream to switch the version to 0.18[0] however
it seems that this is still pending, so we can pin it for now until
that merges.

[0]: https://github.com/fabric8io/docker-maven-plugin/pull/770

Change-Id: I0ed6216a76058a27133ffbab4af50cb8b8bbcd9a
Issue-ID: CIMAN-154
Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>