Updated git submodules
Project: appc master 39db412ce7214d74e0df638c3dc4368bc0a55456

Revert query changes

The changes between patch 3 and patch 4 of this commit
https://gerrit.onap.org/r/#/c/55093, where the parameters
are sent to the SvcLogicContext instead of values, breaks
the save functionality in cdt. When SvcLogicContext is
allowed to subsitute the values, the StringUtils.isEmpty
functions are skipped over.

Although it appears this change is changing a parameterized
query to a static statement query, in fact the query is not
parameterized in its current form either. In its current form,
SvcLogicContext is just subsituting the values into a static
statement query. SvcLogicContext does perform a simple single
quote escape, which could be duplicated here.

Change-Id: If3ffceed5ad147c402355ffef17a8abf0752d5b1
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-ID: APPC-1107