Updated git submodules
Project: so master 8ad7083991ee9e9a0c026028cebc6ab24ca4d272

Docker naming convention

SO is not following the naming convention.  Docker image names
should have an "onap" prefix.  The image names are changed as
follows:

so/catalog-db-adapter --> onap/so/catalog-db-adapter
so/request-db-adapter --> onap/so/request-db-adapter
so/sdnc-adapter --> onap/so/sdnc-adapter
so/openstack-adapter --> onap/so/openstack-adapter
so/vfc-adapter --> onap/so/vfc-adapter
so/asdc-controller --> onap/so/asdc-controller
so/bpmn-infra --> onap/so/bpmn-infra
so/api-handler-infra --> onap/so/api-handler-infra

A corresponding change is needed in INT's docker-manifest.csv.

Change-Id: I024e42607d5fc57a712ad67709f241ee8720a670
Issue-ID: SO-870
Signed-off-by: Rob Daugherty <rd472p@att.com>