Updated git submodules
Project: so master 4d4177a26a5ccbead0da4818cbf93be6aeb39617

Merge "openstack_adapter container fails to start"

openstack_adapter container fails to start

The reported problem occurs because the openstack-adapters app
is attempting to do a migration on top of a migration already
performed by the catalog-db app.

I'm adding these options to flyway in openstack-adapters:
  baseline-on-migrate: true
  validate-on-migrate: false

Another issue is that the java-based CloudConfig migration class
was looking for CloudConfig data only in the application.yaml
file in the classpath.  It was not looking for CloudConfig data
in the override file.  I've changed this logic to look in the
override file first and then in application.yaml.  Ideally, these
sources would be merged, but I don't see a reasonable way to do it.

Change-Id: I7ba07c1f8f00b4c628e825393ee31502950fe592
Issue-ID: SO-868
Signed-off-by: Rob Daugherty <rd472p@att.com>