Updated git submodules
Project: integration master 09a544c2d95de2649ec30c5e521d3b503a5fe478

Merge "Change URI to point to onap git"

Change URI to point to onap git

The gerrit link which was used before is not working as expected.
The downloaded file was with a different name which is difficult
to parse. So the URL is now changed to onap git reference.

Change-Id: I4c4cb289d40eb5340f502c027f4e900ca422e6ba
Issue-ID: INT-626
Signed-off-by: Sudhakar Reddy <Sudhakar.Reddy@amdocs.com>