Updated git submodules
Project: oom master 38a4f2cc12dff3072cdfa777155f5f43107b6170

Merge "Sync docker tags with release manifest"

Sync docker tags with release manifest

Sync docker tags with the release manifest.  It is
assumed that the release manifest is the
"single source of truth" and that the versions in
the helm charts are out of date.

Change-Id: I2902e60b25f57409dd616780a626e63c92d26769
Issue-ID: INT-663
Signed-off-by: Gary Wu <gary.i.wu@huawei.com>