Updated git submodules
Project: policy/engine master 01a357b4e27480844686eb83ef81809a28056e41

Fix random failures in ManualClientEndTest

Don't know if this is the cause of the intermittent failures, but it
could be, and it is certainly a bug: the classes manipulate data, that
isn't marked volatile, from more than one thread at a time.
Updated licenses.

Change-Id: Ifca84cf8f915ec703a9ce8980db2475dacf97980
Issue-ID: POLICY-1092
Signed-off-by: Jim Hahn <jrh3@att.com>