Update patch set 1
Patch Set 1:

We were using repolist to have at least one reference to a project when first integrated. Since you are already referenced in the developing guide it is not necessary. With the new sphinx_rtd_theme and expandable navigation, it is better to have only one reference and we have been removing/commenting out references in the repolist.   Multiple references to the same index can create unexpected context changes in the expandable navigation.   You/we may want to add references to other sub index or files(eg. release notes, APIs depending on content you are providing.

Patch-set: 1