Removal of index and search from the landing page
Index: The index did not cover the whole documentation
and the index entries were not referenced from the text,
therefore it is not useful.
The long term plan is to replace it with a proper glossary,
so the '.. index::' tags are not removed.

Search: The search page is an empty page, while the left bar
already contains a working search function.

Issue-ID: DOC-356
Change-Id: Ibda434a302d90ed16d4f8c2c5b986d16306f0c25
Signed-off-by: Gergely Csatari <gergely.csatari@nokia.com>