Updated git submodules
Project: multicloud/k8s master 073d2f7806257f27b5b92088dc59ae5a98ae6c38

Merge "Delete content when a definition is deleted"

Delete content when a definition is deleted

When a definition is deleted, delete the content too.
Not deleting the content would leave blobs of data that is not
connected to any metadata in the mongo database.

Issue-ID: MULTICLOUD-393
Change-Id: Ifae55bab0affc22b06b1069a27cd90af6f7d5d98
Signed-off-by: Kiran Kamineni <kiran.k.kamineni@intel.com>