Update patch set 1
Patch Set 1:

Yes this was the intention. The sequence in which merges are made to the doc project and contributing submodule - or - a submodule branch name defined in gerrit but without changes made to the branch can result in a reference in the doc project that is not tracking the HEAD of the submodule branch.

Patch-set: 1