Update patch set 1
Patch Set 1:

It is broken because of git metadata - the exact reason why in the doc page is said: Wait for the above change to be merged before any merge to the project repository that you have just added as a submodule. If the project repository added as submodule changes before the doc project merge, git may not automatically update the submodule reference on changes and/or the verify job will fail in the step below.
https://docs.onap.org/en/casablanca/guides/onap-developer/how-to-use-docs/include-documentation.html#setup-project-repositories-s

And the reason why we did the freeze - this can be fixed I guess only by removing the submodule and let them do it again correctly

Patch-set: 1