Update patch set 2
Patch Set 2:

Hi @Manish Kumar, thanks for that - it directed me to the problematic argument "--depth" - which did some shallow update and it break with that ccsdk/cds submodule. I made a new patchset to deal with this better than in the previous version.

Unfortunately if someone wants to build the doc it must be patient to download very slowly (at least here in Europe) 2,4G of repo data (crazy).

Patch-set: 2