Update patch set 2
Patch Set 2:

With --depth 1 it is much more faster process and does not spend so much time and space with downloading 2,4+GB of nonsense (vs 700MB with --depth 1). But to make it work I have to remove that broken submodule ccsdk/cds - so that probably would be better change then these two my patchsets so far.

I am puzzled that no one cares - master is broken, the working instructions of Manish Kumar are working in the same way as my second patchset - but it so slow and such a waste of storage.

How does the community build the docs? The official procedure is broken, that is why I have these patchsets here, but original tox.ini seems to be fine - you just must not merge something what breaks master and if you do, fix it right away by reverting it.

I would like @RichBennett to tell me what he thinks.

Thanks

Patch-set: 2