Update patch set 2
Patch Set 2:

@Petr Ospaly - 

I found the issue causing the tox -elocal to fail.  The doc project references to ccsdk/apps.git, ccsdk/cds.git, oom/offline-installer.git as submodules are not the HEAD of master branch.  Without a --depth 1 option the git submodule update succeeds and with a --depth 1 it fails.   When it succeeds you get older versions of those three submodules published HTML.

I have not had time to confirm the exact cause, but the order of the initial commits on doc repo vs submodule repos could cause this.   There was also a gerrit upgrade that impacted the tracking of submodules for a few weeks of February https://gerrit.onap.org/r/#/c/78448/

I agree shallow clones are more efficient, especially on the ~2 ONAP repos that are large.
I wanted to use depth 1 in the verify job, but Jenkins SCM and/or the JJB scripts do not support the submodule update depth 1.   

I think we want to leave the tox -elocal logic as is detecting a failure on a local test.   

You should continue to use depth 1 for local tests.  

Perhaps we should add a daily documentation test that performs more extensive tests that are expensive or can't be done in a verify job (check HEAD of branch references, validity of URLs, summarizes RST warnings by project, etc).

I will submit a patch to advance the 3 submodule references and add you as reviewer.

Patch-set: 2
Reviewer: Gerrit User 19 <19@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0