Create policy doc structure
The Policy Framework is being re-built, so docs folders need
to be created for the new components. The old architecture will
remain through Frankfurt, but parts of that documentation need
to be reorganized.

policy/parent repo will house the overall documentation, release
notes, etc. It will point to relevant repo where more specifics
will be documented.

Change-Id: I162fdb0420d82eb7347248e4151b9b504adb5817
Issue-ID: POLICY-1677
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>