Structural changes to documentation
Link corrected in landing page to the overview document.
We moved out from the Developers Guide: Architecture, Developing ONAP, Setting-up ONAP.
We renamed 'Developing ONAP' section to 'ONAP components and functionalities'. (Later
we plan to further improve the current content of the Developing ONAP section, that is,
we e.g. plan to move 'Documentation' part to Developers Guide).
We are not 100% sure that 'ARchitecture'/'Setting up ONAP'/'ONAP components and functionalities'
(current Developing ONAP) shall be 'duplicated or not, that is, shall be visible on
the left bar AND under Developers Guide as well, OR only in the left bar.

Issue-ID: DOC-449

Change-Id: Idca0bacac3c7cb28a870a7ea5f8beb54fda0187c
Signed-off-by: Noemi Wagner <noemi.wagner@nokia.com>