Link correction in Landing page and removal of Dublin release naming
The link to VNF/PNF requirements doc has been corrected in the table
on the landing page. Also the name 'Dublin release' has been removed
from the text. The final sentence on navigation has been removed
as it referred to the left bar, which might change in the future
with the new theme.

Issue-ID: DOC-353

Change-Id: Ic0b8b861dd83a338caa5403ccef9175d1af6ea43
Signed-off-by: Noemi Wagner <noemi.wagner@nokia.com>