Remove trailing slash in osa submodule
Gerrit is not properly updating doc repo when we merge patches
in osa repo.

My current theory is that gerrit cannot properly match
a submodule to gerrit project as in .gitmodules project name is
written with trailing slash. As it's the only submodule that has this
trailing slash let's give it a try and remove that slash.

Issue-ID: SECCOM-246
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: I578432a6518c94ad8911202542a9a037acc60db5
(cherry picked from commit bf2a337cff9ce673cae3808c5591090f88932a54)