Document temporary character of references in repolist.rst
According to the comment in repolist.rst references there are only
temporary and should be removed when the project is referenced from
more appropriate place. Unfortunately developer may never notice this
comment as writting to this file in documentation is done by echo
command. To improve this situation let's add an explicit mention about
this temporary character of repolist.rst to the documentation.

Issue-ID: DOC-501
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: I8d8b5fe03d29b5141d57e1d06220bd6382a52861