Committer promotion - Andreas Geissler
Nominating Andreas as a committer to the doc project.
Andreas has been an active part since the start of the Dublin release
and has contributed a lot. Both in participating in meetings, work sessions and
overlooking the structure of the documentation as well as the division of
content between the wiki and docs.onap.org

Issue-ID: DOC-544

Change-Id: I832277d0d15aa185b33fa415f5f01ec7855e4db2
Signed-off-by: Sofia Wallin <sofia.wallin@est.tech>