Fix automated updates of submodules in repo.
Since:
1afb8c8248 ("Rewrite relative urls in .gitmodules")
95d09016c1 ("Fix the last relative urls")

automated updates of submodules in doc repo are broken because gerrit
tracks only submodules that are hosted on the same gerrit server.
According to gerrit doc:

"Gerrit will only automatically update superprojects where the submodules
are hosted on the same Gerrit instance as the superproject. Gerrit
determines this by checking that the URL of the submodule specified in
the .gitmodules file starts with gerrit.canonicalWebUrl. The protocol
part is ignored in this check."

To fix this issue let's update all the submodules urls to use
gerrit.onap.org instead of git.onap.org.

Issue-ID: DOC-552
Fixes: 1afb8c8248 ("Rewrite relative urls in .gitmodules")
Fixes: 95d09016c1 ("Fix the last relative urls")
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: I98043b3372badf583f1ceab600cf59eeee0cd2e7