Update patch set 4
Patch Set 4:

(1 comment)

Yes, for the sphinxcontrib-needs extension to work, there needs to be a template in the conf.py file that signifies the metadata allowed. This was removed to the conf.py file from the Doc project and not added in the onap.yaml file you pointed to. Not sure if other projects use the sphinxcontrib-needs extension but this is necessary for the vnfrqts project.

I commented on the section that is needed in the conf.py file.

Patch-set: 4
Reviewer: Gerrit User 3271 <3271@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0