Update patch set 3
Patch Set 3:

What we are trying to achieve is to provide generic templates for all project. However, projects will have different needs and we encourage everyone to edit as needed. Add headings for specific needs or remove heading when not applicable. I'd be happy to specify this further for each section.

Patch-set: 3
Label: Code-Review=0