Update patch set 1
Patch Set 1:

Hello,
We have an intern who just arrived in our team.
I wanted to explain ONAP and thought that Frankfurt architecture page (https://docs.onap.org/en/frankfurt/guides/onap-developer/architecture/onap-architecture.html) would help but this page in the current layout (so before this review) is very hard to use:

* ONAP HLD is rather OK, although it's hard to tell which components is used for instanciation for example

* Figure 2 numbers are not aligned with numbers on top:
  For example we have a big "1" on figure 2 for TMF alignment and sentence on 1. is "Design time environment for onboarding services and resources into ONAP and designing required services."
  It's very confusing and we don't really know why we have these two things stitched together...

* Then we move to MicroService part and we go back to Portal / Design time / instanciation. I believe Micro    Service support should be put at the end (or removed, as it's like for an auto company to be very proud to use wheels IMHO)

* No Schema / links on "Design time", just a lot of sentences where it's hard to capture what's interesting and what's not. The one in https://onap-doc.readthedocs.io/en/latest/guides/onap-user/design/index.html is nice but as there are no link between these pages, it can be very hard to find.

* No Schema / links on "Runtime Framework" (by the way, why one is XXX time and the other is YYY framework?). Compared to Design time, There are no schema also on user guide so I guess we're loosing 90% of the people here...

* Closed loop automation has a (very) generic schema but we don't know which components are involved

* We show blueprints but no help / links on how to instantiate them. So, how to do when we want to use "5G blueprint" for example?

To say the least, my personal opinion is that this part of documentation is not good and as it's one the front door of ONAP, it won't help to make the community bigger...

Changes in this review are just "cleaning" a little bit the page but the overall layout is left intact and then this page will still be very poor.
I'm not putting "-1" because I believe it's too late by I think it desserves it.

Patch-set: 1
CC: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>