Update patch set 3
Patch Set 3:

> Patch Set 1: Verified+1
> 
> Great. Looks better.
> 2 questions: 
> why SDNR and ORAN are not under CCSDK table ?
The same with CDS:
I wanted to have them a bit more presented as a component under a sepaarate chapter, otherwise you need to know, that they are belonging to CCSDK
> Why Integration and Modeling are not listed in alphabetical order ?
I wanted also to destinguish between modules/components (as shown in the architecture picture) and projects/functions. Therefor I separated them and put to the top of the page.

But I am open for any other suggestions.
This will be anyway only a draft version, as not all component content is clarified, so we can discuss this also later (after my next week's vacation)

Patch-set: 3