Update patch set 2
Patch Set 2:

(1 comment)

I have also raised some Tickets for the release Notes in MultiCloud and OOM, as they are not matching the templates and raised some question to CCSDK and Release Management about the "Maintenance" projects.
So I would like to keep this patchset open untinn we have a clarification.

Patch-set: 2