Update patch set 1
Patch Set 1: Code-Review+1

Checked the so-bpmn-infra configMap in ONAP/Guilin (SO 1.7.10 containers):
(...)   
    tenant:
      endpoint: http://so-openstack-adapter.onap:8087/services/TenantAdapter
    vnf:
      endpoint: http://so-openstack-adapter.onap:8087/services/VnfAdapter
      rest:
        endpoint: http://so-openstack-adapter.onap:8087/services/rest/v2/vnfs
    volume-groups:
      rest:
        endpoint: http://so-openstack-adapter.onap:8087/services/rest/v1/volume-groups
(...)

This means, VNFAdapter APi v2 is already included as a basic SO-BPMN-INFRA configuration in Guilin.
No reason to mention this in the operator-guide for Guilin, this check makes the procedure more complex, than what is actually is.

Patch-set: 1
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1