Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review+2
> > 
> > @Guillaume: OK
> 
> @Guillaume: Thank you for all your good work! Do you suggest that we open a ticket for the integration project to let them fix the file? By the way: i did not find and objects.inv in the integration repo?!

The object.inv is generated by sphinx and publicly available at this URL
https://docs.onap.org/projects/onap-integration/en/latest/objects.inv
It is compressed in a non-standard way with zlib, you need 'sphobjinv' python tool to decode it.

Up to you to open a ticket, I did not manage to clearly identify if the problem was on the IT support side or on a project side.

Patch-set: 3
Reviewer: Gerrit User 3685 <3685@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>