CI: Add noop verify job in docs
Our submit rules requires 2 +1 Verify votes. One by GHA user and
another by CI or a committer. We need to provide a +1 vote to allow
changes to be submittable.

For now, this action is just a no operation action. If in future an
additional verification is needed in the docs repo, it will replace
this no operation.

Issue-ID: CIMAN-33
Change-Id: I0c1b56d7bbaa97707e967cdc84e85a09748edad2
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>